pyproject.toml
[tool.poetry]
name = "pyralkit"
version = "0.1.0"
description = ""
authors = ["melody"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.1"
dacite = "^1.6.0"
python-dateutil = "^2.8.2"
aiolimiter = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
